Add dialog on trashActivity

+++ app/src/main/java/com/showmiso/swipecontacts/TrashActivity.kt
@@ -3,7 +3,9 @@ package com.showmiso.swipecontacts
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.showmiso.swipecontacts.dialog.TwoButtonDialogFragment
 import com.showmiso.swipecontacts.model.Contact
 import com.showmiso.swipecontacts.presenter.ContactPresenter
 import kotlinx.android.synthetic.main.activity_trash.*
@@ -18,6 +20,9 @@ class TrashActivity : AppCompatActivity(), TrashAdapter.OnItemClickListener {
             this
         )
     }
+    private val trashAdapter by lazy {
+        TrashAdapter(this)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -31,18 +36,30 @@ class TrashActivity : AppCompatActivity(), TrashAdapter.OnItemClickListener {
     }
 
     private fun initUI() {
-        val trashAdapter = TrashAdapter(this)
         list_contact.adapter = trashAdapter
         list_contact.layoutManager = LinearLayoutManager(this)
+        val itemDecorator = DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
+        getDrawable(R.drawable.bg_divider)?.let { itemDecorator.setDrawable(it) }
+        list_contact.addItemDecoration(itemDecorator)
         trashAdapter.updateContact(deleteList)
-
         btn_delete_all.setOnClickListener(onClickListener)
+        btn_back.setOnClickListener(onClickListener)
     }
 
     private val onClickListener = View.OnClickListener {
         when (it.id) {
             R.id.btn_delete_all -> {
-                contactPresenter.deleteContactList(deleteList)
+                val dialog = TwoButtonDialogFragment(
+                    getString(R.string.trash_delete_all),
+                    getString(R.string.dialog_message),
+                    object : TwoButtonDialogFragment.OnDialogClickListener {
+                        override fun clickOk() {
+                            contactPresenter.deleteContactList(deleteList)
+                            trashAdapter.deleteAll()
+                        }
+                    }
+                )
+                dialog.show(supportFragmentManager, "")
             }
             R.id.btn_back -> {
                 onBackPressed()
@@ -50,7 +67,10 @@ class TrashActivity : AppCompatActivity(), TrashAdapter.OnItemClickListener {
         }
     }
 
-    override fun onDeleteItem(contact: Contact) {
-        contactPresenter.deleteContact(contact)
+    override fun onClickItem(contact: Contact, layoutId: Int) {
+        if (layoutId == R.id.btn_delete) {
+            contactPresenter.deleteContact(contact)
+        }
+        deleteList.remove(contact)
     }
 }
